+++ UserGraph.DataLayer.Interfaces/IUsersRepository.cs
@@ -10,9 +10,9 @@ namespace UserGraph.DataLayer.Interfaces
         Task<User> CreateUser(User user);
 
         Task<User[]> GetFollowers(string userId);
-        Task<int> GetFollowersCount(string userId);
+        Task<long> GetFollowersCount(string userId);
         Task<User[]> GetFollowing(string userId);
-        Task<int> GetFollowingCount(string userId);
+        Task<long> GetFollowingCount(string userId);
         Task Follow(string sourceUserId, string destinationUserId);
         Task Unfollow(string sourceId, string destinationId);
     }
